+++ src/Inline/Parser/BacktickParser.php
@@ -19,7 +19,7 @@ use League\CommonMark\Inline\Element\Text;
 use League\CommonMark\InlineParserContext;
 use League\CommonMark\Util\RegexHelper;
 
-class BacktickParser extends AbstractInlineParser
+class BacktickParser implements InlineParserInterface
 {
     /**
      * @return string[]