+++ src/Inline/Parser/InlineParserInterface.php
@@ -16,15 +16,6 @@ use League\CommonMark\InlineParserContext;
 interface InlineParserInterface
 {
     /**
-     * Get the name of the parser
-     *
-     * Note that this must be unique with its block type.
-     *
-     * @return string
-     */
-    public function getName();
-
-    /**
      * @return string[]
      */
     public function getCharacters();