+++ src/Util/RegexHelper.php
@@ -154,27 +154,6 @@ final class RegexHelper
     const REGEX_LINK_DESTINATION_BRACES = '/^(?:' . '[<](?:[^ <>\\t\\n\\\\\\x00]' . '|' . self::PARTIAL_ESCAPED_CHAR . '|' . '\\\\)*[>]' . ')/';
 
     /**
-     * @deprecated Instance methods will be removed in 0.18 or 1.0 (whichever comes first)
-     */
-    protected static $instance;
-
-    /**
-     * @return RegexHelper
-     *
-     * @deprecated Instances are no longer needed and will be removed in 0.18 or 1.0
-     */
-    public static function getInstance()
-    {
-        @trigger_error('RegexHelper no longer uses the singleton pattern. Directly grab the REGEX_ or PARTIAL_ constant you need instead.', E_USER_DEPRECATED);
-
-        if (self::$instance === null) {
-            self::$instance = new self();
-        }
-
-        return self::$instance;
-    }
-
-    /**
      * @param string|null $character
      *
      * @return bool
@@ -186,100 +165,6 @@ final class RegexHelper
         }
 
         return preg_match('/' . self::PARTIAL_ESCAPABLE . '/', $character) === 1;
-    }
-
-    /**
-     * Returns a partial regex
-     *
-     * It'll need to be wrapped with /.../ before use
-     *
-     * @param int $const
-     *
-     * @return string
-     *
-     * @deprecated Just grab the constant directly
-     */
-    public function getPartialRegex($const)
-    {
-        @trigger_error('RegexHelper no longer supports the getPartialRegex() function. Directly grab the PARTIAL_ constant you need instead.', E_USER_DEPRECATED);
-
-        switch ($const) {
-            case self::ESCAPABLE: return self::PARTIAL_ESCAPABLE;
-            case self::ESCAPED_CHAR: return self::PARTIAL_ESCAPED_CHAR;
-            case self::IN_DOUBLE_QUOTES: return self::PARTIAL_IN_DOUBLE_QUOTES;
-            case self::IN_SINGLE_QUOTES: return self::PARTIAL_IN_SINGLE_QUOTES;
-            case self::IN_PARENS: return self::PARTIAL_IN_PARENS;
-            case self::REG_CHAR: return self::PARTIAL_REG_CHAR;
-            case self::IN_PARENS_NOSP: return self::PARTIAL_IN_PARENS_NOSP;
-            case self::TAGNAME: return self::PARTIAL_TAGNAME;
-            case self::BLOCKTAGNAME: return self::PARTIAL_BLOCKTAGNAME;
-            case self::ATTRIBUTENAME: return self::PARTIAL_ATTRIBUTENAME;
-            case self::UNQUOTEDVALUE: return self::PARTIAL_UNQUOTEDVALUE;
-            case self::SINGLEQUOTEDVALUE: return self::PARTIAL_SINGLEQUOTEDVALUE;
-            case self::DOUBLEQUOTEDVALUE: return self::PARTIAL_DOUBLEQUOTEDVALUE;
-            case self::ATTRIBUTEVALUE: return self::PARTIAL_ATTRIBUTEVALUE;
-            case self::ATTRIBUTEVALUESPEC: return self::PARTIAL_ATTRIBUTEVALUESPEC;
-            case self::ATTRIBUTE: return self::PARTIAL_ATTRIBUTE;
-            case self::OPENTAG: return self::PARTIAL_OPENTAG;
-            case self::CLOSETAG: return self::PARTIAL_CLOSETAG;
-            case self::OPENBLOCKTAG: return self::PARTIAL_OPENBLOCKTAG;
-            case self::CLOSEBLOCKTAG: return self::PARTIAL_CLOSEBLOCKTAG;
-            case self::HTMLCOMMENT: return self::PARTIAL_HTMLCOMMENT;
-            case self::PROCESSINGINSTRUCTION: return self::PARTIAL_PROCESSINGINSTRUCTION;
-            case self::DECLARATION: return self::PARTIAL_DECLARATION;
-            case self::CDATA: return self::PARTIAL_CDATA;
-            case self::HTMLTAG: return self::PARTIAL_HTMLTAG;
-            case self::HTMLBLOCKOPEN: return self::PARTIAL_HTMLBLOCKOPEN;
-            case self::LINK_TITLE: return self::PARTIAL_LINK_TITLE;
-        }
-    }
-
-    /**
-     * @return string
-     *
-     * @deprecated Use PARTIAL_HTMLTAG and wrap it yourself instead
-     */
-    public function getHtmlTagRegex()
-    {
-        @trigger_error('RegexHelper::getHtmlTagRegex() has been deprecated. Use the RegexHelper::PARTIAL_HTMLTAG constant instead.', E_USER_DEPRECATED);
-
-        return '/^' . self::PARTIAL_HTMLTAG . '/i';
-    }
-
-    /**
-     * @return string
-     *
-     * @deprecated Use PARTIAL_LINK_TITLE and wrap it yourself instead
-     */
-    public function getLinkTitleRegex()
-    {
-        @trigger_error('RegexHelper::getLinkTitleRegex() has been deprecated. Use the RegexHelper::PARTIAL_LINK_TITLE constant instead.', E_USER_DEPRECATED);
-
-        return '/' . self::PARTIAL_LINK_TITLE . '/';
-    }
-
-    /**
-     * @return string
-     *
-     * @deprecated Use REGEX_LINK_DESTINATION_BRACES instead
-     */
-    public function getLinkDestinationBracesRegex()
-    {
-        @trigger_error('RegexHelper::getLinkDestinationBracesRegex() has been deprecated. Use the RegexHelper::REGEX_LINK_DESTINATION_BRACES constant instead.', E_USER_DEPRECATED);
-
-        return self::REGEX_LINK_DESTINATION_BRACES;
-    }
-
-    /**
-     * @return string
-     *
-     * @deprecated Use the REGEX_THEMATIC_BREAK constant directly
-     */
-    public function getThematicBreakRegex()
-    {
-        @trigger_error('RegexHelper::getThematicBreakRegex() has been deprecated. Use the RegexHelper::REGEX_THEMATIC_BREAK constant instead.', E_USER_DEPRECATED);
-
-        return self::REGEX_THEMATIC_BREAK;
     }
 
     /**